backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
onnxruntime==1.20.1
Pillow==11.0.0
numpy==2.3.0
python-multipart==0.0.6
